% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_skip.R
\name{coverage_skip}
\alias{coverage_skip}
\title{Run covr with failing tests}
\usage{
coverage_skip(test_path = "tests/testthat", test_skip = NULL)
}
\arguments{
\item{test_path}{character, path to test directory, Default: 'tests/testthat'}

\item{test_skip}{data.frame, mapping of failing tests, PARAM_DESCRIPTION, Default: NULL}
}
\description{
Run \code{\link[covr]{package_coverage}} ignoring failing tests from \code{\link[testthat]{test_dir}}.
}
\details{
If test_skip is NULL then a new test_dir will be run internally checking
for failing tests.
}
\seealso{
\code{\link[testthat]{test_dir}}
 \code{\link[covr]{package_coverage}}

Other utility: \code{\link{covr_summary}},
  \code{\link{map_testthat_snapshot}},
  \code{\link{map_testthat}}, \code{\link{map_test}},
  \code{\link{testthat_summary}}
}
\concept{utility}
