% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covrpage_log.R
\name{covr_log}
\alias{covr_log}
\title{Query 'git log' for code coverage history}
\usage{
covr_log()
}
\value{
data.frame
}
\description{
Retrieve log of code coverages evalulated by covrpage that
were commited to git version control.
}
\examples{
\dontrun{
covr_log()
}
}
\seealso{
Other log: \code{\link{covrpage_log}}
}
\concept{log}
