% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_chunks.R
\name{covr_summary}
\alias{covr_summary}
\title{FUNCTION_TITLE}
\usage{
covr_summary(x)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{mutate}}
}
