% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covrpage.R
\name{covrpage}
\alias{covrpage}
\title{Render covrpage}
\usage{
covrpage(path = getwd(), preview = TRUE, auto_push = FALSE,
  update_badge = TRUE, vignette = FALSE)
}
\arguments{
\item{path}{path to package}

\item{preview}{boolean, to open the output in viewer, Default: TRUE}

\item{auto_push}{boolean, push to remote repo on exit, Default: FALSE}

\item{update_badge}{boolean, locate badge in README and update with
testthat results, Default: TRUE}

\item{vignette}{boolean, if TRUE then the covrpage README will be copied
as a vignette of the package, Default: FALSE}
}
\description{
Render the template of the covrpage in the package directory.
}
\seealso{
\code{\link[rmarkdown]{render}}

Other invoke: \code{\link{covrpage_ci}},
  \code{\link{covrpage_snapshot}}
}
\concept{invoke}
