% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covrpage_ci.R
\name{covrpage_ci}
\alias{covrpage_ci}
\title{covrpage for travis}
\usage{
covrpage_ci(path = getwd(), ...)
}
\arguments{
\item{path}{character, path to package, Default: getwd()}

\item{\dots}{arguments to pass to covrpage}
}
\description{
wrapper for covrpage call with switches for running
after-success on travis and deploying gh-pages.
}
\examples{
\dontrun{
if(interactive()){
 covrpage_ci()
 }
}
}
