% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covrpage_log.R
\name{covrpage_log}
\alias{covrpage_log}
\title{Query 'git log' for covrpage history}
\usage{
covrpage_log()
}
\value{
list
}
\description{
Retrieve log of covrpage README.md files committed to git
version control
}
\examples{
\dontrun{
covrpage_log()
}
}
\seealso{
Other log: \code{\link{covr_log}}
}
\concept{log}
