% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covrpage_snapshot.R
\name{covrpage_snapshot}
\alias{covrpage_snapshot}
\title{Run covrpage on a GitHub Repository}
\usage{
covrpage_snapshot(repo)
}
\arguments{
\item{repo}{character, Public GitHub repository}
}
\description{
Runs a covrpage on a GitHub repository in a sterile environment
}
\details{
Creates a markdown output that is sent to the internal viewer.
}
\examples{
\dontrun{
if(interactive()){
 covrpage_snapshot('tidyverse/purrr')
 }
}
}
