% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_testthat.R
\name{map_test}
\alias{map_test}
\title{Map a single test file}
\usage{
map_test(path)
}
\arguments{
\item{path}{character, path to file}
}
\value{
list
}
\description{
Return nested list of context/test/expectation in testthat R file
}
\seealso{
\code{\link[utils]{getParseData}}
}
