% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_testthat.R
\name{map_testthat}
\alias{map_testthat}
\title{Hierarchy structure of testthat tests}
\usage{
map_testthat(path = "tests/testthat")
}
\arguments{
\item{path}{character, path to tests, Default: 'tests/testthat'}
}
\value{
list
}
\description{
Return nested list of context/test/expectation in testthat directory
}
\seealso{
\code{\link[stats]{setNames}}
}
