% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{package_name}
\alias{package_name}
\title{package name}
\usage{
package_name(path = getwd())
}
\arguments{
\item{path}{path to package}
}
\description{
Retrieve package name
}
\seealso{
Other utility: 
\code{\link{coverage_skip}()},
\code{\link{covr_summary}()},
\code{\link{map_testthat_snapshot}()},
\code{\link{map_testthat}()},
\code{\link{map_test}()},
\code{\link{testthat_summary}()}
}
\concept{utility}
