% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tencrypt.R
\name{tencrypt}
\alias{tencrypt}
\title{encrypt travis environment variables}
\usage{
tencrypt(r_obj = Sys.getenv("GITHUB_PAT"), travis_env = "GH_PAT",
  add = TRUE)
}
\arguments{
\item{r_obj}{object value to encrypt}

\item{travis_env}{character, name of travis environment variable}

\item{add}{boolean, add directly to .travis.yml, Default: TRUE}
}
\description{
add encrypted travis environment variable to git directory and .travis.yml.
}
