% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_covrpage.R
\name{use_covrpage}
\alias{use_covrpage}
\title{Add covrpage to Travis CI for a git directory}
\usage{
use_covrpage(file = ".travis/covrpage.sh", gh_user = gsub("@(.*?)$", "",
  system("git config user.email", intern = TRUE)), push_branch = "test",
  repo = system("git config travis.slug", intern = TRUE),
  deploy_branch = system("git rev-parse --abbrev-ref HEAD", intern = TRUE))
}
\arguments{
\item{file}{character, file path to save bash script to, Default: '.travis/covrpage.sh'}

\item{gh_user}{character, Github user name,
Default: gsub("@(.*?)$", "", system("git config user.email", intern = TRUE))}

\item{push_branch}{character, branch Travis will create to push updated covrpage readme file, Default: 'test'}

\item{repo}{character, name of git repo that package is hosted on, Default: system("git config travis.slug", intern = TRUE)}

\item{deploy_branch}{character, name of branch that current package build is located,
Default: system("git rev-parse --abbrev-ref HEAD", intern = TRUE) (usually 'master')}
}
\description{
Adds bash file to execute after Travis test is successfully completed.
}
