% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vignette.R
\name{use_covrpage_vignette}
\alias{use_covrpage_vignette}
\title{Create a vignette for covrpage}
\usage{
use_covrpage_vignette(path = ".")
}
\arguments{
\item{path}{character, path to root of file, Default:'.'}
}
\description{
Perform general setup and copying of tests/README.md into
vignette subdirectory
}
\examples{
\dontrun{
use_covrpage_vignette()
}
}
\seealso{
Other setup: 
\code{\link{make_badge}()},
\code{\link{tencrypt}()},
\code{\link{use_covrpage}()}
}
\concept{setup}
