% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_chunks.R
\name{covr_summary}
\alias{covr_summary}
\title{Summary outputs for covr object}
\usage{
covr_summary(x, package, failed = FALSE)
}
\arguments{
\item{x}{covr object}

\item{package}{character, package name}

\item{failed}{flag for failed test, Default: FALSE}
}
\value{
data.frame
}
\description{
Creates a summary data.frame for covr object output that
mimics the console print method of \code{\link[covr]{package_coverage}}
}
\seealso{
\code{\link[covr]{package_coverage}}

Other utility: 
\code{\link{coverage_skip}()},
\code{\link{map_testthat_snapshot}()},
\code{\link{map_testthat}()},
\code{\link{map_test}()},
\code{\link{testthat_summary}()}
}
\concept{utility}
