% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_chunks.R
\name{covr_summary}
\alias{covr_summary}
\title{Summary outputs for covr object}
\usage{
covr_summary(x)
}
\arguments{
\item{x}{covr object}
}
\value{
data.frame
}
\description{
Creates a summary data.frame for covr object output that 
mimics the console print method of \code{\link[covr]{package_coverage}}
}
\seealso{
\code{\link[covr]{package_coverage}}
}
