% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covrpage.R
\name{covrpage}
\alias{covrpage}
\title{render covrpage}
\usage{
covrpage(pkg, add_lines = TRUE, preview = TRUE, auto_push = FALSE)
}
\arguments{
\item{pkg}{path to package}

\item{preview}{boolean, to open the output in viewer, Default: TRUE}

\item{auto_push}{boolean, push to remote repo on exit, Default: FALSE}
}
\description{
Render the template of the covrpage in the package directory.
}
\seealso{
\code{\link[rmarkdown]{render}}
}
