% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covrpage_snapshot.R
\name{covrpage_snapshot}
\alias{covrpage_snapshot}
\title{Run covrpage on a GitHub Repository}
\usage{
covrpage_snapshot(
  repo,
  branch = NULL,
  PAT = Sys.getenv("GITHUB_PAT"),
  host = "api.github.com",
  preview = TRUE
)
}
\arguments{
\item{repo}{character, Public GitHub repository}

\item{branch}{character, Name of the branch to checkout. Default: NULL}

\item{PAT}{character, GitHub Personal Access Token, Default: Sys.getenv("GITHUB_PAT")}

\item{host}{character, GitHub API host to use, Default: "api.github.com"}

\item{preview}{boolean, to open the output in viewer, Default: TRUE}
}
\description{
Runs a covrpage on a GitHub repository in a sterile environment
}
\details{
Creates a markdown output that is sent to the internal viewer.

Setting branch to NULL means that \code{\link[git2r]{clone}} will use the remote's default branch.
}
\examples{
\dontrun{
if(interactive()){
 covrpage_snapshot('tidyverse/purrr')
 }
}
}
\seealso{
\code{\link[covrpage]{covrpage}} \code{\link[git2r]{clone}} \code{\link[remotes]{install_github}}

Other invoke: 
\code{\link{covrpage_ci}()},
\code{\link{covrpage}()}
}
\concept{invoke}
