% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge.R
\name{make_badge}
\alias{make_badge}
\title{Create covrpage Badge}
\usage{
make_badge(remote_origin = NULL, active_branch = NULL,
  tinyurl = TRUE)
}
\arguments{
\item{remote_origin}{character, username/repo of the remote origin, Default: NULL}

\item{active_branch}{character, current active git branch, Default: NULL}

\item{tinyurl}{boolean, convert badge uri to tinyurl link, Default: TRUE}
}
\value{
character
}
\description{
Create covrpage badge for test status and date that is connected to
the README.md created in the tests directory.
}
\details{
if either inputs are NULL then the .git file is inspected to gather the 
information
}
