% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_testthat.R
\name{map_test}
\alias{map_test}
\title{Map the hierarchy structure of a testthat test file}
\usage{
map_test(path)
}
\arguments{
\item{path}{character, path to file}
}
\value{
data.frame
}
\description{
Query a test file to see what expectations are defined in it.
}
\details{
Return data.frame containing which expectations are in the test file
 by context/description/test/expectation/linerange
}
\seealso{
\code{\link[utils]{getParseData}}

Other utility: \code{\link{coverage_skip}},
  \code{\link{covr_summary}}, \code{\link{map_testthat}},
  \code{\link{testthat_summary}}
}
\concept{utility}
