% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tencrypt.R
\name{tencrypt}
\alias{tencrypt}
\title{Encrypt Travis environment variables}
\usage{
tencrypt(
  r_obj = Sys.getenv("GITHUB_PAT"),
  travis_env = "GITHUB_PAT",
  add = TRUE
)
}
\arguments{
\item{r_obj}{object value to encrypt}

\item{travis_env}{character, name of Travis environment variable}

\item{add}{boolean, add directly to .travis.yml, Default: TRUE}
}
\description{
Add encrypted Travis environment variable to git directory and .travis.yml.
}
\seealso{
Other setup: 
\code{\link{make_badge}()},
\code{\link{use_covrpage_vignette}()},
\code{\link{use_covrpage}()}
}
\concept{setup}
