% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_chunks.R
\name{testthat_summary}
\alias{testthat_summary}
\title{Summary outputs for testthat object}
\usage{
testthat_summary(x, type = c("short", "long"))
}
\arguments{
\item{x}{list, testthat object output}

\item{type}{character, format of the object output, Default: c("short", "long")}
}
\value{
data.frame
}
\description{
summary outputs for testthat object that is used in rmd
}
